% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.R
\name{findActiveParam}
\alias{findActiveParam}
\title{Finds the active model's parameters}
\usage{
findActiveParam(M, n, k, na.rm = TRUE)
}
\arguments{
\item{M}{matrix}

\item{n}{number of units (equal to number of \code{M}'s rows)}

\item{k}{parameters to retrieve}

\item{na.rm}{logical, whether to ignore \code{NA} data}
}
\value{
An array containing the parameters
}
\description{
Finds the active model's parameters
}
