% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochBlock-package.R
\docType{package}
\name{StochBlock-package}
\alias{StochBlock}
\alias{StochBlock-package}
\title{StochBlock: Stochastic Blockmodeling of One-Mode and Linked Networks}
\description{
Stochastic blockmodeling of one-mode and linked networks as implemented in Škulj and Žiberna (2022) <10.1016/j.socnet.2022.02.001>.
}
\author{
\strong{Maintainer}: Aleš Žiberna \email{ales.ziberna@fdv.uni-lj.si} (\href{https://orcid.org/0000-0003-1534-6971}{ORCID})

Other contributors:
\itemize{
  \item Fabio Ashtar Telarico \email{Fabio-Ashtar.Telarico@fdv.uni-lj.si} (\href{https://orcid.org/0000-0002-8740-7078}{ORCID}) [contributor]
}

}
\keyword{internal}
